requirements_pipeline.txt
fastapi
uvicorn
sse-starlette
groq
requests
PyGithub
python-dotenv
httpx
flask
